split region out of config.rs
