+++ shipcat_definitions/src/merge.rs
@@ -1,6 +1,6 @@
 // This file describes how manifests and environment manifest overrides are merged.
 
-use config::{Config, Region};
+use super::{Config, Region};
 use super::{Manifest, Result};
 
 impl Manifest {
